move helper functions to utils module
